docs: add PieChartWithNeedle example in TypeScript
